requirements.txt
customtkinter~=5.0.3
spotipy~=2.22.0
python-dotenv~=0.21.0
future~=0.18.2
pygame~=2.1.2
requests~=2.28.1
Pillow~=9.3.0